pyproject.toml
[project]
name = "babysitter-poc"
version = "0.1.0"
description = "Autonomous Claude loop with human-in-the-loop via Notion"
requires-python = ">=3.10"
dependencies = [
    "notion-client>=2.0.0",
    "mcp>=0.1.0",
    "tiktoken>=0.12.0",
    "claude-agent-sdk>=0.1.10",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "fastapi>=0.122.0",
    "uvicorn>=0.38.0",
    "python-dotenv>=1.2.1",
    "websockets>=15.0.1",
    "httpx>=0.28.1",
]

[tool.hatch.build.targets.wheel]
packages = ["scripts", "src"]

[tool.pytest.ini_options]
pythonpath = "."
